requirements.txt
anthropic>=0.41.0
langchain==0.3.13
langchain-anthropic==0.3.3
faiss-cpu==1.9.0
sentence-transformers==3.3.1
pydantic==2.10.3
python-dotenv==1.0.1
numpy==1.26.4
streamlit==1.41.1
jupyter==1.1.1
pytest==8.3.4
